Move authentication services to its folder
